requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
colorama==0.4.6
python-dotenv==1.0.1
lxml==4.9.3

# Required by pandas
numpy>=1.26.2
python-dateutil>=2.8.2
pytz>=2023.3.post1
tzdata>=2023.3

# Required by requests
certifi>=2023.11.17
charset-normalizer>=3.3.2
idna>=3.6
urllib3>=2.1.0

# Required by beautifulsoup4
soupsieve>=2.5

# Standard library modules used (no installation needed)
# - argparse
# - datetime
# - json
# - logging
# - os
# - re
# - sys
# - time
